refactor: moved UnitExtensions to core
